+++ smarts/core/default_map_builder.py
@@ -24,30 +24,24 @@ from typing import NamedTuple, Tuple
 from smarts.core.road_map import RoadMap
 from smarts.core.utils.file import file_md5_hash
 
-
-class _RoadMapInfo(NamedTuple):
-    source: str
-    lanepoint_spacing: float
-    default_lane_width: float
-    obj: RoadMap
-    map_hash: str
-
-
 _existing_map = None
 
 
+# This is an instance of a sstudio.types.MapBuilder Callable
+# which should return an object derived from the RoadMap base class
+# and a hash that uniquely identifies it (changes to the hash should signify
+# that the map is different enough that map-related caches should be reloaded).
+#
+# This function should be re-callable (although caching is up to the implementation).
 # The idea here is that anything in SMARTS that needs to use a RoadMap
-# can call this factory to get or create one of default type.
+# can call this builder to get or create one of default type.
 #
 # Downstream developers who want to extend SMARTS to support other
-# map formats (by extending the RoadMap base class) can replace this
-# file with their own version and shouldn't have to change much else.
-def get_road_map(
-    map_source: str,
-    lanepoint_spacing: float = None,
-    default_lane_width: float = None,
-    no_cache: bool = False,
-) -> Tuple[RoadMap, str]:
+# map formats (by extending the RoadMap base class) can create their
+# own version of this to reference from a MapSpec within their
+# scenario folder(s) and shouldn't have to change much else.
+# TODO:  update this when OpenDrive support is ready, to choose format based on file extension
+def get_road_map(map_spec, no_cache: bool = False) -> Tuple[RoadMap, str]:
     """@return a RoadMap object and a hash
     that uniquely identifies it. Changes to the hash
     should signify that the map is different enough
@@ -57,14 +51,12 @@ def get_road_map(
     unless the no_cache parameter is True, in which case
     a new object will always be created.
     """
-
-    assert map_source, "A road map source must be specified"
+    assert map_spec, "A road map spec must be specified"
+    assert map_spec.source, "A road map source must be specified"
 
     global _existing_map
     if _existing_map:
-        if not no_cache and _existing_map.obj.is_same_map(
-            map_source, lanepoint_spacing, default_lane_width
-        ):
+        if not no_cache and _existing_map.obj.is_same_map(map_spec):
             return _existing_map.obj, _existing_map.map_hash
         import gc
 
@@ -73,11 +65,13 @@ def get_road_map(
         _existing_map = None
         gc.collect()
 
-    map_path = map_source
+    map_path = map_spec.source
     if not os.path.isfile(map_path):
-        map_path = os.path.join(map_source, "map.net.xml")
+        map_path = os.path.join(map_spec.source, "map.net.xml")
         if not os.path.exists(map_path):
-            raise FileNotFoundError(f"Unable to find map in map_source={map_source}.")
+            raise FileNotFoundError(
+                f"Unable to find map in map_source={map_spec.source}."
+            )
 
     # Keep this a conditional import so Sumo does not have to be
     # imported if not necessary:
@@ -85,15 +79,20 @@ def get_road_map(
 
     road_map = SumoRoadNetwork.from_file(
         map_path,
-        default_lane_width=default_lane_width,
-        lanepoint_spacing=lanepoint_spacing,
+        default_lane_width=map_spec.default_lane_width,
+        lanepoint_spacing=map_spec.lanepoint_spacing,
     )
 
     road_map_hash = file_md5_hash(road_map.source)
 
     if not no_cache:
-        _existing_map = _RoadMapInfo(
-            map_source, lanepoint_spacing, default_lane_width, road_map, road_map_hash
-        )
+        from smarts.sstudio.types import MapSpec
+
+        class _RoadMapInfo(NamedTuple):
+            map_spec: MapSpec
+            obj: RoadMap
+            map_hash: str
+
+        _existing_map = _RoadMapInfo(map_spec, road_map, road_map_hash)
 
     return road_map, road_map_hash